requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
extruct==0.18.0
firebase-admin==6.8.0
Flask==3.1.0
flask-cors==5.0.1
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
html5lib==1.1
html_text==0.7.0
httplib2==0.22.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jstyleson==0.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mf2py==2.0.1
msgpack==1.1.0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
pyRdfa3==3.6.4
python-dotenv==1.1.0
rdflib==7.1.4
recipe_scrapers==15.7.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
w3lib==2.3.1
webencodings==0.5.1
Werkzeug==3.1.3
